feat(orders): purchase/sales order returns with replace/refund types and reopening
Adds a full return lifecycle for both order types: starting a return
immediately creates and locks a credit-note invoice, completing it
reverses inventory/product stock, and voiding a pending return
releases the reserved quantity. A return can be typed 'replace'
(leaves a real gap requiring re-receiving/re-delivery) or 'refund'
(settles the quantity via a new quantity_refunded column so the order
can complete without further receiving/delivery). Starting any return
on an already-complete order reopens it immediately, requiring an
explicit re-complete afterward.

Also widens sales-order quantity columns (quantity_ordered/delivered/
returned/refunded, delivery and return line items) from Integer to
Numeric(15,2) to match purchase orders' existing precision.

+++ app/models/sales_delivery_item.py
@@ -1,5 +1,5 @@
 """Sales delivery item model - line items in each delivery"""
-from sqlalchemy import Column, Integer, ForeignKey, Text
+from sqlalchemy import Column, Integer, ForeignKey, Text, Numeric
 from sqlalchemy.orm import relationship
 from app.db.base_class import Base
 
@@ -22,7 +22,7 @@ class SalesDeliveryItem(Base):
     item_id = Column(Integer, ForeignKey("items.id", ondelete="RESTRICT"), nullable=True, index=True)
 
     # === QUANTITY ===
-    quantity_delivered = Column(Integer, nullable=False)  # Quantity in THIS delivery
+    quantity_delivered = Column(Numeric(15, 2), nullable=False)  # Quantity in THIS delivery
 
     # === NOTES ===
     notes = Column(Text, nullable=True)
